cooldown, extinguisher preserves second wind

+++ code/game/objects/effects/aliens.dm
@@ -307,6 +307,9 @@
 	var/barricade_damage = 40
 	var/in_weather = FALSE
 
+	/// Set when attempting to clear acid off of an item with extinguish_acid() to prevent an item being extinguished multiple times in a tick.
+	COOLDOWN_DECLARE(clear_acid)
+
 //Sentinel weakest acid
 /obj/effect/xenomorph/acid/weak
 	name = "weak acid"
@@ -409,9 +412,18 @@
 
 /obj/effect/xenomorph/acid/proc/finish_melting()
 	playsound(src, "acid_hit", 25, TRUE)
-	visible_message(SPAN_XENODANGER("[acid_t] collapses under its own weight into a puddle of goop and undigested debris!"))
+
+	if(istype(acid_t, /obj/item/weapon/gun))
+		var/obj/item/weapon/gun/acid_gun = acid_t
+		if(acid_gun.has_second_wind)
+			visible_message(SPAN_XENODANGER("[acid_t] loses its shine as the acid bubbles against it."))
+			acid_gun.has_second_wind = FALSE
+			playsound(src, 'sound/weapons/handling/gun_jam_click.ogg', 25, TRUE)
+			qdel(src)
+			return
 
 	if(istype(acid_t, /turf))
+		visible_message(SPAN_XENODANGER("[acid_t] is terribly damaged by the acid covering it!"))
 		if(istype(acid_t, /turf/closed/wall))
 			var/turf/closed/wall/wall = acid_t
 			new /obj/effect/acid_hole(wall)
@@ -421,33 +433,39 @@
 
 	else if (istype(acid_t, /obj/structure/girder))
 		var/obj/structure/girder/girder = acid_t
+		visible_message(SPAN_XENODANGER("[acid_t] collapses and falls in on itself as the acid melts its frame!"))
 		girder.dismantle()
 
 	else if(istype(acid_t, /obj/structure/window/framed))
 		var/obj/structure/window/framed/window = acid_t
+		visible_message(SPAN_XENODANGER("[acid_t] audibly cracks and fails as the acid bubbles against it!"))
 		window.deconstruct(disassembled = FALSE)
 
 	else if(istype(acid_t, /obj/structure/barricade))
+		visible_message(SPAN_XENODANGER("[acid_t] cracks and fragments as the acid sizzles against it!"))
 		pass() // Don't delete it, just damaj
 
 	else
 		for(var/mob/mob in acid_t)
 			mob.forceMove(loc)
+		visible_message(SPAN_XENODANGER("[acid_t] collapses under its own weight into a puddle of goop and undigested debris!"))
 		qdel(acid_t)
 	qdel(src)
 
 /obj/effect/xenomorph/acid/extinguish_acid()
+	if(!COOLDOWN_FINISHED(src, clear_acid))
+		return
+	COOLDOWN_START(src, clear_acid, 1 SECONDS)
+
 	if(istype(acid_t, /obj/item/weapon/gun))
 		var/obj/item/weapon/gun/acid_gun = acid_t
-		if(acid_gun.has_second_wind)
-			acid_gun.has_second_wind = FALSE
-			visible_message(SPAN_XENODANGER("[acid_t] loses its shine as the acid is sprayed off of it!"))
-			playsound(src, 'sound/weapons/handling/gun_jam_click.ogg', 25, TRUE)
+		if(!acid_gun.has_second_wind)
+			visible_message(SPAN_XENODANGER("[acid_t] seems unaffected and continues to deform!"))
+			return FALSE
+		else
+			visible_message(SPAN_XENODANGER("The sizzling on [acid_t] quiets as the acid is sprayed off of it!"))
 			qdel(src)
 			return TRUE
-		else
-			visible_message(SPAN_XENODANGER("[acid_t] seems unaffected and continues to deform."))
-	return FALSE
 
 /obj/effect/xenomorph/boiler_bombard
 	name = "???"
